Changed estimatedPrice type to int (originally double) This is because the field is called estimated price This also keeps the interface simple and clean without any decimal places There is also no need for formatting The max length of the field has been set to 9 characters, which is meant to not have Integer.parseInt not throw a number format exception

+++ app/src/main/java/behrman/justin/shoppinglist/dialog/NewShoppingItemDialog.java
@@ -73,7 +73,7 @@ public class NewShoppingItemDialog {
         String estimatedCostString = costText.getText().toString().trim();
         String description = descriptionText.getText().toString().trim();
         if (invalidFields(itemName, estimatedCostString, description)) return;
-        double estimatedCost = Double.parseDouble(estimatedCostString);
+        int estimatedCost = Integer.parseInt(estimatedCostString);
         boolean purchased = purchasedBox.isChecked();
         ShoppingItem item = new ShoppingItem(itemName, description, estimatedCost, (Category) categorySpinner.getSelectedItem(), purchased);
         this.onFinish.accept(item);
